+++ src/yafowil/plone/connectors.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+
+from collections.abc import MutableMapping
 from zope.i18n import translate
 from zope.i18nmessageid import Message
 from ZPublisher.HTTPRequest import FileUpload