+++ src/yafowil/plone/resources.py
@@ -94,13 +94,7 @@ class YafowilJS(Resources):
 
     def get_resources_content(self, resources):
         data = StringIO()
-        data.write('var mockup_require = require;\n')
-        data.write('var mockup_define = define;\n')
-        data.write('require = undefined;\n')
-        data.write('define = undefined;\n')
         self.read_resources_content(data, resources)
-        data.write('require = mockup_require;\n')
-        data.write('define = mockup_define;\n')
         return data.getvalue()